BF - A bit of refactoring and completiong of SLA and SLT

+++ 1130.SystemObjects/Instructions/AddDouble.cs
@@ -9,7 +9,7 @@
 		{
 			var effectiveAddress = GetEffectiveAddress(state);
 			long valueToAdd = (state[effectiveAddress] << 16) & 0xffff0000;
-			valueToAdd |= state[effectiveAddress | 1] & 0xffff;
+			valueToAdd |= (ushort) (state[effectiveAddress | 1] & 0xffff);
 			long accExt = state.AccExt;
 			long result = accExt + valueToAdd;
 			state.AccExt = (uint) (result & 0xffffffff);
